+++ src/report/pull-request-report-findings-filter.ts
@@ -1,18 +1,18 @@
 import { GitHub } from '@actions/github/lib/utils'
 import { ExtendedContext } from '../github/extended-context'
 import { SecHubCodeFinding, SecHubFinding } from '../model/sechub'
-import { PullRequestCompareMode } from '../input/inputs'
+import { PullRequestFilterMode } from '../input/inputs'
 import { ReportFindingsFilter } from './report-findings-filter'
 
 export class PullRequestReportFindingsFilter implements ReportFindingsFilter {
   constructor(
     private readonly octokit: InstanceType<typeof GitHub>,
     private readonly context: ExtendedContext,
-    private readonly pullRequestCompareMode: PullRequestCompareMode
+    private readonly pullRequestFilterMode: PullRequestFilterMode
   ) {}
 
   async filter(findings: SecHubFinding[]): Promise<SecHubFinding[]> {
-    if (this.pullRequestCompareMode === PullRequestCompareMode.NONE)
+    if (this.pullRequestFilterMode === PullRequestFilterMode.NONE)
       return findings
 
     const { data } = await this.octokit.rest.pulls.listFiles({
@@ -43,11 +43,11 @@ export class PullRequestReportFindingsFilter implements ReportFindingsFilter {
     let found = false
     do {
       found = currentCodeFinding.location === filename
-      switch (this.pullRequestCompareMode) {
-        case PullRequestCompareMode.ENTRY_POINT:
+      switch (this.pullRequestFilterMode) {
+        case PullRequestFilterMode.ENTRY_POINT:
           currentCodeFinding = undefined
           break
-        case PullRequestCompareMode.CALL_HIERARCHY:
+        case PullRequestFilterMode.CALL_HIERARCHY:
           currentCodeFinding = currentCodeFinding.calls
           break
       }