+++ src/report/report-properties.ts
@@ -1,7 +1,4 @@
-import { PullRequestCompareMode } from '../input/inputs'
-
 export interface ReportProperties {
   maxSize?: number
   considerErrorOnSeverities: string[]
-  pullRequestCompareMode: PullRequestCompareMode
 }